% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{binomial_sim}
\alias{binomial_sim}
\title{Run a binomial simulation with a specific sample sieze, probability,
and number of iterations.}
\usage{
binomial_sim(iter = 100, n = 10, prob = 0.7)
}
\arguments{
\item{iter}{provide the number of iterations you wish to run in your
simulation.}

\item{n}{provide the size of the sample you want per iteration.}

\item{prob}{provide the probability of 1 occurring in your simulation.}
}
\value{
nothing, just print out a barplot and the resulting tables of
the simulation.
}
\description{
Run a binomial simulation with a specific sample sieze, probability,
and number of iterations.
}
\examples{
binomial_sim(iter = 10000, n = 5, prob = 0.2)
binomial_sim(iter = 500, n = 10, prob = 0.8)
}
