% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt.R
\name{clt_binom}
\alias{clt_binom}
\title{Run a binomial distribution simulation of n samples iter times.
Print out the resulting histogram superimposed with a theoretical
normal curve and corresponding density curve.}
\usage{
clt_binom(n, iter, p = 0.5)
}
\arguments{
\item{n}{provide the size of the sample you want per iteration.}

\item{iter}{provide the number of iterations to experiment with.}

\item{p}{the probability for success for any given sample.}
}
\value{
nothing, just print out resulting histogram.
}
\description{
Run a binomial distribution simulation of n samples iter times.
Print out the resulting histogram superimposed with a theoretical
normal curve and corresponding density curve.
}
\examples{
clt_binom(n = 5, iter = 10000, p = 0.7)
}
