% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt.R
\name{clt_pois}
\alias{clt_pois}
\title{Run a poisson distribution simulation of n samples iter times.
Print out the resulting histogram superimposed with a theoretical
normal curve and corresponding density curve.}
\usage{
clt_pois(n, iter, lambda = 10)
}
\arguments{
\item{n}{provide the size of the sample you want per iteration.}

\item{iter}{provide the number of iterations to experiment with.}

\item{lambda}{provide the average number of successes for the
underlying Poisson distribution samples.}
}
\value{
nothing, just print out resulting histogram.
}
\description{
Run a poisson distribution simulation of n samples iter times.
Print out the resulting histogram superimposed with a theoretical
normal curve and corresponding density curve.
}
\examples{
clt_pois(n = 5, iter = 10000, lambda = 5)
}
