% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt.R
\name{clt_unif}
\alias{clt_unif}
\title{Run a uniform distribution simulation of n samples iter times.
Print out the resulting histogram superimposed with a theoretical
normal curve and corresponding density curve.}
\usage{
clt_unif(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{provide the size of the sample you want per iteration.}

\item{iter}{provide the number of iterations to experiment with.}

\item{a}{provide the lower bound for samples.}

\item{b}{provide the upper bound for samples.}
}
\value{
nothing, just print out resulting histogram.
}
\description{
Run a uniform distribution simulation of n samples iter times.
Print out the resulting histogram superimposed with a theoretical
normal curve and corresponding density curve.
}
\examples{
clt_unif(n = 5, iter = 10000, a = 0, b = 5)
}
