% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bayes}
\alias{get_bayes}
\title{Returns the probability of event A given B with Baye's
rule.}
\usage{
get_bayes(a, b, ba)
}
\arguments{
\item{a}{The probability of event A happening.}

\item{b}{The probability of event B happening.}

\item{ba}{The probability of event B given event A.}
}
\value{
A given B.
}
\description{
Returns the probability of event A given B with Baye's
rule.
}
\examples{
p <- get_bayes(0.75, 0.6, 0.25)
}
