% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bayes_testing}
\alias{get_bayes_testing}
\title{Run Baye's testing theorem on given probabilities.}
\usage{
get_bayes_testing(u, tu, tubar)
}
\arguments{
\item{u}{The probability someone is a user}

\item{tu}{The probability someone tests positive
given they are a user. More specifically, what is the
probability for a true positive.}

\item{tubar}{The probability someone tests positive given
that they aren't a user. More specifically, what is the
probability someone gets a false positive when they aren't
a user.}
}
\value{
The probability someone is a user given they test
positive.
}
\description{
Run Baye's testing theorem on given probabilities.
}
\examples{
pct <- get_bayes_testing(3, 95, 30)
}
