% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_chebyshev}
\alias{get_chebyshev}
\title{Get the minimum percentage of observations within k
standard deviations of the mean for any given sample.}
\usage{
get_chebyshev(k)
}
\arguments{
\item{k}{provide the number of standard deviations from the
mean you are hoping to capture.}
}
\value{
The percentage of sample within k standard deviations
of the mean.
}
\description{
Get the minimum percentage of observations within k
standard deviations of the mean for any given sample.
}
\examples{
pct <- get_chebyshev(2)
pct <- get_chebyshev(4)
}
