% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mss}
\alias{get_mss}
\title{Get the mean sum of squares for a vector of quantitative
data as predicted values.}
\usage{
get_mss(y, yhat)
}
\arguments{
\item{y}{provide the vector of actual values you are wishing
to model against.}

\item{yhat}{provide the vector of fitted values for the dataset
based on your model.}
}
\value{
The MSS of a dataset
}
\description{
Get the mean sum of squares for a vector of quantitative
data as predicted values.
}
\examples{
mss <- get_mss(c(1, 2, 3, 4, 5), c(2, 3, 4, 5, 6))
}
