% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_permutations}
\alias{get_permutations}
\title{Execute the permutations rule with a set of nset
elements, and removal or n elements in distinct ordering.}
\usage{
get_permutations(nset, n)
}
\arguments{
\item{nset}{The number of elements within the set.}

\item{n}{The number of elements you want to take out
of the set in a distinct ordering.}
}
\value{
Permutations of nset / n.
}
\description{
Execute the permutations rule with a set of nset
elements, and removal or n elements in distinct ordering.
}
\examples{
p <- get_permutations(30, 10)
}
