% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_t_statistic}
\alias{get_t_statistic}
\title{Execute the permutations rule with a set of nset
elements, and removal or n elements in distinct ordering.}
\usage{
get_t_statistic(x, mu)
}
\arguments{
\item{x}{Vector of input data to compute statistic on.}

\item{mu}{The actual mean of the population.
of the set in a distinct ordering.}
}
\value{
Permutations of nset / n.
}
\description{
Returns the number of permutations given two numbers.
Function implements the following calculation:
nset! / (nset - n)!
}
\examples{
p <- get_permutations(30, 10)
}
