% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_with_margins}
\alias{get_table_with_margins}
\title{Create a table with marginals given two variables.}
\usage{
get_table_with_margins(x, y)
}
\arguments{
\item{x}{The x param to utilize in your table.}

\item{y}{The y param to utilize in your table.}
}
\value{
The table printed to stdout.
}
\description{
Create a table with marginals given two variables.
}
\examples{
get_table_with_margins(c(1,2,3,4), c(4,5,6,7))
}
