% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_variance}
\alias{get_variance}
\title{Get the variance for a vector of quantitative data.}
\usage{
get_variance(vector)
}
\arguments{
\item{vector}{provide the vector you wish to get the variance from}
}
\value{
The variance of a dataset
}
\description{
Get the variance for a vector of quantitative data.
}
\examples{
s_squared <- get_variance(c(1, 2, 3, 4, 5))
}
