% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_z}
\alias{get_z}
\title{Get the z values for a vector}
\usage{
get_z(vector)
}
\arguments{
\item{vector}{provide the vector you wish to extract z values from}
}
\value{
another vector with z values
}
\description{
Get the z values for a vector
}
\examples{
z <- get_z(c(1, 2, 3, 4, 5))
}
