% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project1.R
\name{ntickets}
\alias{ntickets}
\title{Predict number of tickets to sell for a flight}
\usage{
ntickets(n = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{n}{provide the number of seats available on a flight.}

\item{gamma}{provide the probability of overbooking a flight.}

\item{p}{provide the probability of someone showing up for a flight.}
}
\value{
list of solution, N, gamma, and p
}
\description{
Predict the number of tickets you are able to sell given the
number of seats available and the probability of overbooking/someone
actually showing up. Utilizes both binomial and continuous normal
distributions for predicting natural and real numbers for the
number of tickets to sell.
}
\examples{
ntickets()
ntickets(n = 400)
}
