% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{get_norm_curve}
\alias{get_norm_curve}
\title{Plots a normal curve distribution}
\usage{
get_norm_curve(mu = 0, sigma = 10)
}
\arguments{
\item{mu}{provide the mean of the standard distribution you wish to model.}

\item{sigma}{provide the standard deviation of the distribution you wish
to model.}
}
\value{
nothing, just print out a barplot and the resulting tables of
the simulation.
}
\description{
Plots a normal curve distribution
}
\examples{
get_norm_curve(mu = 5, sigma = 3)
get_norm_curve(mu = 10, sigma = 5)
}
